+++ Sources/Client/Model/Message/Message.swift
@@ -28,6 +28,7 @@ public struct Message: Codable {
         case latestReactions = "latest_reactions"
         case ownReactions = "own_reactions"
         case reactionScores = "reaction_scores"
+        case isSilent = "silent"
     }
     
     /// A custom extra data type for messages.
@@ -72,6 +73,8 @@ public struct Message: Codable {
     public private(set) var ownReactions: [Reaction]
     /// A reactions count (see `ReactionCounts`).
     public private(set) var reactionScores: [String: Int]
+    /// Flag for silent messages. Silent messages won't increase the unread count. See https://getstream.io/chat/docs/silent_messages/?language=swift
+    public var isSilent: Bool
     
     /// Check if the message is ephemeral, e.g. Giphy preview.
     public var isEphemeral: Bool { type == .ephemeral }
@@ -114,6 +117,7 @@ public struct Message: Codable {
     ///   - updated: Message updated date
     ///   - deleted: Message deleted date
     ///   - text: Message body
+    ///   - isSilent: Flag for indicating if this message is silent (does not increase unread count)
     ///   - command: Command associated with message
     ///   - args: Arguments in message
     ///   - attachments: Attachments for this message
@@ -131,6 +135,7 @@ public struct Message: Codable {
                 updated: Date = .init(),
                 deleted: Date? = nil,
                 text: String,
+                silent: Bool = false,
                 command: String? = nil,
                 args: String? = nil,
                 attachments: [Attachment] = [],
@@ -149,6 +154,7 @@ public struct Message: Codable {
         self.updated = updated
         self.deleted = deleted
         self.text = text
+        self.isSilent = silent
         self.command = command
         self.args = args
         self.attachments = attachments
@@ -164,6 +170,7 @@ public struct Message: Codable {
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(text, forKey: .text)
+        try container.encode(isSilent, forKey: .isSilent)
         extraData?.encodeSafely(to: encoder, logMessage: "📦 when encoding a message extra data")
         
         if !attachments.isEmpty {
@@ -189,6 +196,7 @@ public struct Message: Codable {
         updated = try container.decode(Date.self, forKey: .updated)
         deleted = try container.decodeIfPresent(Date.self, forKey: .deleted)
         text = try container.decode(String.self, forKey: .text).trimmingCharacters(in: .whitespacesAndNewlines)
+        isSilent = try container.decodeIfPresent(Bool.self, forKey: .isSilent) ?? false
         command = try container.decodeIfPresent(String.self, forKey: .command)
         args = try container.decodeIfPresent(String.self, forKey: .args)
         attachments = try container.decode([Attachment].self, forKey: .attachments)